Version 3.1
Finished doing logging system. Now when properties of object are
updated, or properties of collections inside object are changed, state
of this object is logged. If object has a key that mathches
"ClassName"+"ID" template, this key will be also logged as additional
information.  If we delete object its state also logged.
I added error page, which is shown by filter if exception arises, this
filter also logs  exeption with message, stack trace and optional
description.