Improved AJAX requests
Improved AJAX requests. Now we send to the server not the title of
university/faculty/speciality, but its id. As a response we get JSON
object wich contains array of objects, each object at least has:  object
ID in database and and its name, title and all information that can be
efficiently used in view.
Also made dropdowns for choosing county and city for entrant.